Update tests for CommandLineApplication class
